requirements.txt
google-api-python-client
google-auth-oauthlib
google-auth-httplib2
groq
edge-tts
moviepy
Pillow
python-dotenv
requests